pyproject.toml
[tool.poetry]
name = "generate_flashcards"
version = "1.0.0"
description = "Using Python, GPT4 and LangChain to Generate Custom Anki Decks"
authors = ["Orest Tokovenko <hi@oresttokovenko.com>"]
readme = "README.md"
license = "MIT"
packages = [
    {include = "src"}
]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.12"
langchain_openai = "^0.0.8"
pandas = "^2.2.1"
python-dotenv = "^1.0.1"
genanki = "^0.13.1"
openai = "^1.14.1"
pandas-stubs = "^2.2.1.240316"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.3"
mypy = "^1.9.0"
nox-poetry = "^1.0.3"
poetry-plugin-export = "^1.7.1"
isort = "^5.13.2"

[tool.ruff]
target-version = 'py311'

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.mypy]
python_version = "3.11"
strict = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = "noxfile.py"
follow_imports = "normal"

[tool.poetry.scripts]
generate-flashcards = 'src.generate_flashcards:main'
generate-deck = 'src.generate_deck:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
